1. FG super-channel config\ 2. command parsing modified according to L format

+++ WSS_Main/src/PatternCalibModule/ParametersStructure.h
@@ -12,41 +12,41 @@
 struct Aopt_Kopt_ThreadArgs
 {
 	int 			port;
-	float 			freq;
+	double 			freq;
 
-	float 			result_Aopt;
-	float 			result_Kopt;
+	double 			result_Aopt;
+	double 			result_Kopt;
 };
 
 struct Aatt_Katt_ThreadArgs
 {
 	int 			port;
-	float 			freq;
-	float 			ATT;
+	double 			freq;
+	double 			ATT;
 
-	float 			result_Aatt;
-	float 			result_Katt;
+	double 			result_Aatt;
+	double 			result_Katt;
 };
 
 struct Sigma_ThreadArgs
 {
 	int 			port;
-	float 			freq;
-	float 			temperature;
+	double 			freq;
+	double 			temperature;
 
-	float 			result_Sigma;
+	double 			result_Sigma;
 };
 
 struct Pixel_Pos_ThreadArgs
 {
-	float 			f1;
-	float 			f2;
-	float 			fc;
-	float 			temperature;
+	double 			f1;
+	double 			f2;
+	double 			fc;
+	double 			temperature;
 
-	float 			result_F1_PixelPos;
-	float 			result_F2_PixelPos;
-	float 			result_FC_PixelPos;
+	double 			result_F1_PixelPos;
+	double 			result_F2_PixelPos;
+	double 			result_FC_PixelPos;
 };
 
 #endif /* SRC_PATTERNCALIBMODULE_PARAMETERSSTRUCTURE_H_ */
